Project Introduction
This is my submission for the Shopify Backend Developer Fall 2022 Intern Challenge.  I have created an inventory tracking application for AMB Logistics.  This application is capable of creating inventory, editing, deleting and viewing a list of inventories.  Additionally, this application has the ability to create locations and assign inventory to a specific location.